arreglos

+++ src/controllers/subcategory-controllers.js
@@ -17,14 +17,19 @@ export const ctrlCreateSubcategory = async (req, res) => {
     if (!existingCategory) {
       return res.status(400).json({ error: "La categoría no existe" });
     }
+    console.log("Categoria")
+    console.log(existingCategory)
 
-    // Verifica si ya existe una subcategoría con el mismo nombre
-    const existingSubcategory = await SubcategoryModel.findOne({ subcategory });
+    const existingSubcategory = await SubcategoryModel.findOne({
+      category: existingCategory._id,
+      subcategory,
+    });
     if (existingSubcategory) {
-      return res
-        .status(400)
-        .json({ error: "Ya hay una subcategoría con el mismo nombre" });
+      return res.status(400).json({error: " Ya hay una subcategoria con el mismo nombre"});
     }
+
+    console.log("Subcategoria")
+    console.log(existingSubcategory)
 
     // Crea la nueva subcategoría
     const newSubcategory = new SubcategoryModel({
@@ -32,16 +37,19 @@ export const ctrlCreateSubcategory = async (req, res) => {
       category: existingCategory._id,
       material,
     });
+    console.log("Nueva Categoria")
+    console.log(newSubcategory)
     await newSubcategory.save();
-
-    // Agrega la nueva subcategoría al array de subcategorías de la categoría
+     // Agrega la nueva subcategoría al array de subcategorías de la categoría
     existingCategory.subcategories.push(newSubcategory._id);
     await existingCategory.save();
 
-    res
-      .status(201)
-      .json({ newSubcategory, message: "Subcategoría creada exitosamente" });
-  } catch (error) {
+    res.status(201).json({ newSubcategory, message: "Subcategoría creada exitosamente" });
+  } catch (error)  {
+    // Manejo del error de clave duplicada en el índice compuesto
+    if (error.code === 11000 && error.keyPattern && error.keyPattern.subcategory === 1 && error.keyPattern.category === 1) {
+      return res.status(400).json({ error: "Ya existe una subcategoría con el mismo nombre en esta categoría" });
+    }
     console.error("Error al crear la subcategoría:", error);
     res.status(500).json({ error: "Problema con la base de datos" });
   }
@@ -113,11 +121,21 @@ export const ctrlDeleteSubcategory = async (req, res) => {
 
   try {
     const subcategory = await SubcategoryModel.findById(subcategoryId);
-
+    console.log("Aca en el Delete")
+    console.log(subcategory)
+    console.log(subcategory.category)
     if (!subcategory) {
       return res.status(404).json({ error: "Subcategoría no encontrada" });
     }
 
+    const category = await CategoryModel.findById(subcategory.category);
+    if (!category) {
+      return res.status(404).json({ error: "Categoria no encontrada" });
+    }
+
+    category.subcategories.pull(subcategoryId)
+    await category.save()
+ 
     await SubcategoryModel.deleteOne({ _id: subcategoryId });
     return res
       .status(200)
